Lab8

+++ Lab7/src/routes/book.py
@@ -3,11 +3,18 @@ from flask import Blueprint, jsonify, request
 from src.utils import *
 import src.models as models
 import src.db as db
+from src.routes.user import UserRole
+from src import app
+from flask_jwt_extended import (
+    JWTManager, jwt_required, get_jwt_identity
+)
+jwt = JWTManager(app)
 
 book_bp = Blueprint('book', __name__, url_prefix='/book')
 
 
 @book_bp.route('/', methods=['POST'])
+@jwt_required()
 def create_book():
     class Book(Schema):
         name = fields.Str(required=True)
@@ -20,6 +27,12 @@ def create_book():
         book = Book().load(request.json)
     except ValidationError as error:
         return jsonify(error.messages), 400
+
+    user = db.session.query(models.User).filter(models.User.id == get_jwt_identity()).first()
+    if user is None:
+        return jsonify({'message': 'Unauthorized'}), 404
+    if user.role != UserRole.Administrator:
+        return jsonify({'message': 'Forbidden'}), 401
     if db.session.query(models.Book).filter(models.Book.name == book['name']).count() != 0:
         return jsonify({'message': 'Book already exists'}), 400
     new_book = models.Book(name=book['name'], author_first_name=book['author_first_name'],
@@ -54,6 +67,7 @@ def get_book_status(values):
 
 
 @book_bp.route('/<int:book_id>', methods=['PUT'])
+@jwt_required()
 def update_book(book_id):
     class Book(Schema):
         name = fields.Str(required=True)
@@ -63,6 +77,11 @@ def update_book(book_id):
         book = Book().load(request.json)
     except ValidationError as error:
         return jsonify(error.messages), 400
+    user = db.session.query(models.User).filter(models.User.id == get_jwt_identity()).first()
+    if user is None:
+        return jsonify({'message': 'Unauthorized'}), 404
+    if user.role != UserRole.Administrator:
+        return jsonify({'message': 'Forbidden'}), 401
     book_to_update = db.session.query(models.Book).filter(models.Book.id == book_id).first()
     if book_to_update is None:
         return jsonify({'message': 'Book not found'}), 404
@@ -79,11 +98,20 @@ def update_book(book_id):
 
 
 @book_bp.route('/<int:book_id>', methods=['DELETE'])
+@jwt_required()
 def delete_book(book_id):
     book = db.session.query(models.Book).filter(models.Book.id == book_id).first()
+    user = db.session.query(models.User).filter(models.User.id == get_jwt_identity()).first()
+    orders = db.session.query(models.Order).filter(models.Order.user_id == get_jwt_identity()).all()
+    if user is None:
+        return jsonify({'message': 'Unauthorized'}), 404
+    if user.role != UserRole.Administrator:
+        return jsonify({'message': 'Forbidden'}), 401
     if book is None:
         return jsonify({'message': 'Book not found'}), 404
     try:
+        for order in orders:
+            db.session.delete(order)
         db.session.delete(book)
     except:
         db.session.rollback()
